✨ Make ABI positions and sizes f32 for sub-pixel drawing
Screen-space coordinates and extents (positions, w/h, radius, camera) now
cross the wasm ABI as f32 instead of i32. The host floors each to a pixel
at draw time, matching PICO-8, so carts can carry sub-pixel state for
smooth motion without rounding themselves — and flooring (not truncation)
keeps movement even across x = 0. The wire stays 32-bit, so this costs
nothing on the ABI.

Sprite w/h are f32 too: like PICO-8, a fractional count draws a partial
slice (w = 0.5 is a 4-pixel-wide sprite). Discrete values stay integers:
sprite/tile indices, map cel coordinates, flags, buttons, channels, colors.

Updates the SDK wrappers, raw FFI, host linking, the sprite rasterizer,
the WAT test carts, the scaffolding/import templates, the example carts
(the platformer drops its 1/16-px fixed-point hack and just stores f32
pixels), and the docs. LIMITS.md now points carts at the libm crate for
float math that core lacks (floor/sqrt/abs/trig), since that math runs
cart-side in no_std.

+++ examples/map_demo/src/lib.rs
@@ -6,27 +6,27 @@
 use rico8::*;
 
 struct MapDemo {
-    cam_x: i32,
-    cam_y: i32,
+    cam_x: f32,
+    cam_y: f32,
     solid_only: bool,
 }
 
 impl Game for MapDemo {
     fn update(&mut self, ctx: &mut Context) {
         if ctx.btn(Button::Left) {
-            self.cam_x -= 2;
+            self.cam_x -= 2.0;
         }
         if ctx.btn(Button::Right) {
-            self.cam_x += 2;
+            self.cam_x += 2.0;
         }
         if ctx.btn(Button::Up) {
-            self.cam_y -= 2;
+            self.cam_y -= 2.0;
         }
         if ctx.btn(Button::Down) {
-            self.cam_y += 2;
+            self.cam_y += 2.0;
         }
-        self.cam_x = self.cam_x.clamp(0, 32 * 8 - SCREEN_W);
-        self.cam_y = self.cam_y.clamp(0, 16 * 8 - SCREEN_H + 64);
+        self.cam_x = self.cam_x.clamp(0.0, (32 * 8 - SCREEN_W) as f32);
+        self.cam_y = self.cam_y.clamp(0.0, (16 * 8 - SCREEN_H + 64) as f32);
         if ctx.btnp(Button::X) {
             self.solid_only = !self.solid_only;
         }
@@ -37,17 +37,17 @@ impl Game for MapDemo {
         gfx.camera(self.cam_x, self.cam_y);
         // Layer mask 0 draws everything; mask 1 only flag-0 (solid) tiles.
         let layers = if self.solid_only { 1 } else { 0 };
-        gfx.map(0, 0, 0, 0, 32, 16, layers);
-        gfx.camera(0, 0);
-        gfx.print("arrows scroll, x: layers", 4, 2, Color::WHITE);
+        gfx.map(0, 0, 0.0, 0.0, 32, 16, layers);
+        gfx.camera(0.0, 0.0);
+        gfx.print("arrows scroll, x: layers", 4.0, 2.0, Color::WHITE);
         if self.solid_only {
-            gfx.print("solid tiles only", 4, 120, Color::ORANGE);
+            gfx.print("solid tiles only", 4.0, 120.0, Color::ORANGE);
         }
     }
 }
 
 rico8::game!(MapDemo {
-    cam_x: 0,
-    cam_y: 0,
+    cam_x: 0.0,
+    cam_y: 0.0,
     solid_only: false,
 });
